OperationCenter updated with facade pattern + support mark as done
