+++ Main/test.txt
@@ -1,5 +1,4 @@
-uCode	macroTBeg		macroTEnd		microTBeg	microTEnd	completeStatus	desc
-6	undefined/0/0/0	undefined/0/0/0		-1/-1		-1/-1		false		buy soap
-5	Sunday/22/3/2015	undefined/0/0/0		-1/-1		-1/-1		false		watch football
-1	Sunday/22/3/2015	undefined/0/0/0		-1/-1		-1/-1		false		church
-1	Sunday/22/3/2015	undefined/0/0/0		-1/-1		-1/-1		false		dinner
+6
+uCode	macroTBeg		macroTEnd		microTBeg	microTEnd	completeStatus	priority	alarmMacro		alarmMicro	desc
+6	Monday/23/3/2015		undefined/0/0/0		14/0		-1/-1		false		None		undefined/0/0/0		-1/-1		lunch
+4	Monday/23/3/2015		undefined/0/0/0		18/0		-1/-1		false		None		undefined/0/0/0		-1/-1		dinner
