+++ OperationCenter/OperationCenter.vcxproj
@@ -72,6 +72,7 @@
     <ClInclude Include="OperationCenter.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="Logic.cpp" />
     <ClCompile Include="OperationCenter.cpp" />
   </ItemGroup>
   <ItemGroup>
